Begin migrating base setup to own class.
