+++ tests/tests/bootstrap.php
@@ -16,3 +16,4 @@ if ( ! $_core_dir ) {
 copy( dirname( __FILE__ ) . '/../../object-cache.php', $_core_dir . '/wp-content/object-cache.php' );
 
 require $_tests_dir . '/includes/bootstrap.php';
+require 'base.php';
